+++ package.json
@@ -23,7 +23,7 @@
     "react-redux": "^6.0.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.2",
+    "react-scripts": "2.1.8",
     "redux": "^4.0.1",
     "redux-act": "^1.7.4",
     "redux-logger": "^3.0.6",
