+++ src/components/banner/asset.json
@@ -9,11 +9,6 @@
             "id": 2,
             "url": "https://mp.weixin.qq.com/s/rYlRHY48t8R9SK9nAkZUOA",
             "alt": "Second Banner"
-        },
-        {
-            "id": 3,
-            "url": "https://mp.weixin.qq.com/s/rYlRHY48t8R9SK9nAkZUOA",
-            "alt": "Second Banner"
         }
     ]
 }
