+++ src/components/category-link/index.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Row from "react-bootstrap/lib/Row";
-import Col from "react-bootstrap/lib/Col";
-import Button from "react-bootstrap/lib/Button";
+import Row from "react-bootstrap/Row";
+import Col from "react-bootstrap/Col";
+import Button from "react-bootstrap/Button";
 
 import not_1 from "./IMG/icon-生理问题@3x.png";
 import not_2 from "./IMG/icon-食物介绍@2x.png";
