+++ src/components/course-list/index.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Row from "react-bootstrap/lib/Row";
-import Col from "react-bootstrap/lib/Col";
+import Row from "react-bootstrap/Row";
+import Col from "react-bootstrap/Col";
 import Icon from "../../components/icon";
 
 export default React.memo(({ handler, asset }) => (
