+++ src/components/side-sheet/nav.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "react-bootstrap/lib/Button";
+import Button from "react-bootstrap/Button";
 import { Link } from "react-router-dom";
 import { withRouter } from "react-router";
 import cs from "classnames";
