+++ src/styles/layer/header.scss
@@ -63,7 +63,7 @@
 }
 .header_wiht{
     overflow: hidden;
-    max-width: 130px; 
+    max-width: 140px; 
 }
 .header_detailsle{
     margin-top: 5px;
